requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

requests==2.24.0
psycopg2==2.8.5
PyYAML==5.3.1
Flask==1.1.2
Flask-SocketIO==4.3.1
eventlet==0.29.1
config==0.5.0.post0
rabbitmq==0.2.0
celery==5.0.2
black==20.8b1
mypy==0.790
python-engineio==3.13.1
python-socketio==4.6.0
click==7.1.2

